style: rename files to align the naming convention
